requirements.txt
beautifulsoup4==4.12.3
Brotli==1.1.0
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
cloudscraper==1.2.71
freezegun==1.5.1
html2text==2024.2.26
html5lib==1.1
idna==3.7
parameterized==0.9.0
pillow==10.3.0
pushbullet.py==0.12.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-magic==0.4.27
requests==2.31.0
requests-file==2.0.0
requests-toolbelt==1.0.0
setuptools==58.1.0
six==1.16.0
soupsieve==2.5
urllib3==2.2.1
webencodings==0.5.1
websocket-client==1.8.0
